Split up classes into individual files in the ocr module
